Refactor main.py to unify Menu module into MainMenu.py
